moved constants to new file
