+++ src/main/java/uk/gov/companieshouse/itemhandler/mapper/MapperUtil.java
@@ -4,7 +4,7 @@ import org.mapstruct.Named;
 
 import java.time.LocalDateTime;
 
-import static uk.gov.companieshouse.itemhandler.mapper.OrderDataToOrderConfirmationMapperConstants.DATETIME_OF_PAYMENT_FORMATTER;
+import static uk.gov.companieshouse.itemhandler.util.DateConstants.DATETIME_OF_PAYMENT_FORMATTER;
 
 public interface MapperUtil {
