+++ src/test/java/uk/gov/companieshouse/itemhandler/service/FilingHistoryDescriptionProviderServiceTest.java
@@ -7,7 +7,6 @@ import java.io.File;
 import java.util.HashMap;
 import java.util.Map;
 
-import static java.util.Collections.singletonMap;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.nullValue;
 import static org.hamcrest.core.Is.is;
